+++ spring-web/src/main/java/org/springframework/http/codec/support/BaseDefaultCodecs.java
@@ -1,0 +1,210 @@
+/*
+ * Copyright 2002-2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.http.codec.support;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+import org.springframework.core.codec.ByteArrayDecoder;
+import org.springframework.core.codec.ByteArrayEncoder;
+import org.springframework.core.codec.ByteBufferDecoder;
+import org.springframework.core.codec.ByteBufferEncoder;
+import org.springframework.core.codec.CharSequenceEncoder;
+import org.springframework.core.codec.DataBufferDecoder;
+import org.springframework.core.codec.DataBufferEncoder;
+import org.springframework.core.codec.Decoder;
+import org.springframework.core.codec.Encoder;
+import org.springframework.core.codec.ResourceDecoder;
+import org.springframework.core.codec.StringDecoder;
+import org.springframework.http.codec.CodecConfigurer;
+import org.springframework.http.codec.DecoderHttpMessageReader;
+import org.springframework.http.codec.EncoderHttpMessageWriter;
+import org.springframework.http.codec.FormHttpMessageReader;
+import org.springframework.http.codec.HttpMessageReader;
+import org.springframework.http.codec.HttpMessageWriter;
+import org.springframework.http.codec.ResourceHttpMessageWriter;
+import org.springframework.http.codec.json.Jackson2JsonDecoder;
+import org.springframework.http.codec.json.Jackson2JsonEncoder;
+import org.springframework.http.codec.json.Jackson2SmileDecoder;
+import org.springframework.http.codec.json.Jackson2SmileEncoder;
+import org.springframework.http.codec.xml.Jaxb2XmlDecoder;
+import org.springframework.http.codec.xml.Jaxb2XmlEncoder;
+import org.springframework.lang.Nullable;
+import org.springframework.util.ClassUtils;
+
+/**
+ * Default implementation of {@link CodecConfigurer.DefaultCodecs} that serves
+ * as a base for client and server specific variants.
+ *
+ * @author Rossen Stoyanchev
+ */
+class BaseDefaultCodecs implements CodecConfigurer.DefaultCodecs {
+
+	static final boolean jackson2Present =
+			ClassUtils.isPresent("com.fasterxml.jackson.databind.ObjectMapper",
+					BaseCodecConfigurer.class.getClassLoader()) &&
+					ClassUtils.isPresent("com.fasterxml.jackson.core.JsonGenerator",
+							BaseCodecConfigurer.class.getClassLoader());
+
+	private static final boolean jackson2SmilePresent =
+			ClassUtils.isPresent("com.fasterxml.jackson.dataformat.smile.SmileFactory",
+					BaseCodecConfigurer.class.getClassLoader());
+
+	private static final boolean jaxb2Present =
+			ClassUtils.isPresent("javax.xml.bind.Binder", BaseCodecConfigurer.class.getClassLoader());
+
+
+	@Nullable
+	private Decoder<?> jackson2JsonDecoder;
+
+	@Nullable
+	private Encoder<?> jackson2JsonEncoder;
+
+	private boolean registerDefaults = true;
+
+
+	@Override
+	public void jackson2JsonDecoder(Decoder<?> decoder) {
+		this.jackson2JsonDecoder = decoder;
+	}
+
+	@Override
+	public void jackson2JsonEncoder(Encoder<?> encoder) {
+		this.jackson2JsonEncoder = encoder;
+	}
+
+	/**
+	 * Delegate method used from {@link BaseCodecConfigurer#registerDefaults}.
+	 */
+	void registerDefaults(boolean registerDefaults) {
+		this.registerDefaults = registerDefaults;
+	}
+
+
+	// Package private access to the configured readers...
+
+	final List<HttpMessageReader<?>> getTypedReaders() {
+		if (!this.registerDefaults) {
+			return Collections.emptyList();
+		}
+		List<HttpMessageReader<?>> readers = new ArrayList<>();
+		readers.add(new DecoderHttpMessageReader<>(new ByteArrayDecoder()));
+		readers.add(new DecoderHttpMessageReader<>(new ByteBufferDecoder()));
+		readers.add(new DecoderHttpMessageReader<>(new DataBufferDecoder()));
+		readers.add(new DecoderHttpMessageReader<>(new ResourceDecoder()));
+		readers.add(new DecoderHttpMessageReader<>(StringDecoder.textPlainOnly()));
+		readers.add(new FormHttpMessageReader());
+		extendTypedReaders(readers);
+		return readers;
+	}
+
+	protected void extendTypedReaders(List<HttpMessageReader<?>> typedReaders) {
+	}
+
+	final List<HttpMessageReader<?>> getObjectReaders() {
+		if (!this.registerDefaults) {
+			return Collections.emptyList();
+		}
+		List<HttpMessageReader<?>> readers = new ArrayList<>();
+		if (jackson2Present) {
+			readers.add(new DecoderHttpMessageReader<>(getJackson2JsonDecoder()));
+		}
+		if (jackson2SmilePresent) {
+			readers.add(new DecoderHttpMessageReader<>(new Jackson2SmileDecoder()));
+		}
+		if (jaxb2Present) {
+			readers.add(new DecoderHttpMessageReader<>(new Jaxb2XmlDecoder()));
+		}
+		extendObjectReaders(readers);
+		return readers;
+	}
+
+	protected void extendObjectReaders(List<HttpMessageReader<?>> objectReaders) {
+	}
+
+	final List<HttpMessageReader<?>> getCatchAllReaders() {
+		if (!this.registerDefaults) {
+			return Collections.emptyList();
+		}
+		List<HttpMessageReader<?>> result = new ArrayList<>();
+		result.add(new DecoderHttpMessageReader<>(StringDecoder.allMimeTypes()));
+		return result;
+	}
+
+
+	// Package private access to the configured writers...
+
+	final List<HttpMessageWriter<?>> getTypedWriters() {
+		if (!this.registerDefaults) {
+			return Collections.emptyList();
+		}
+		List<HttpMessageWriter<?>> writers = new ArrayList<>();
+		writers.add(new EncoderHttpMessageWriter<>(new ByteArrayEncoder()));
+		writers.add(new EncoderHttpMessageWriter<>(new ByteBufferEncoder()));
+		writers.add(new EncoderHttpMessageWriter<>(new DataBufferEncoder()));
+		writers.add(new ResourceHttpMessageWriter());
+		writers.add(new EncoderHttpMessageWriter<>(CharSequenceEncoder.textPlainOnly()));
+		extendTypedWriters(writers);
+		return writers;
+	}
+
+	protected void extendTypedWriters(List<HttpMessageWriter<?>> typedWriters) {
+	}
+
+	final List<HttpMessageWriter<?>> getObjectWriters() {
+		if (!this.registerDefaults) {
+			return Collections.emptyList();
+		}
+		List<HttpMessageWriter<?>> writers = new ArrayList<>();
+		if (jackson2Present) {
+			writers.add(new EncoderHttpMessageWriter<>(getJackson2JsonEncoder()));
+		}
+		if (jackson2SmilePresent) {
+			writers.add(new EncoderHttpMessageWriter<>(new Jackson2SmileEncoder()));
+		}
+		if (jaxb2Present) {
+			writers.add(new EncoderHttpMessageWriter<>(new Jaxb2XmlEncoder()));
+		}
+		extendObjectWriters(writers);
+		return writers;
+	}
+
+	protected void extendObjectWriters(List<HttpMessageWriter<?>> objectWriters) {
+	}
+
+	List<HttpMessageWriter<?>> getCatchAllWriters() {
+		if (!this.registerDefaults) {
+			return Collections.emptyList();
+		}
+		List<HttpMessageWriter<?>> result = new ArrayList<>();
+		result.add(new EncoderHttpMessageWriter<>(CharSequenceEncoder.allMimeTypes()));
+		return result;
+	}
+
+
+	// Accessors for use in sub-classes...
+
+	protected Decoder<?> getJackson2JsonDecoder() {
+		return (this.jackson2JsonDecoder != null ? this.jackson2JsonDecoder : new Jackson2JsonDecoder());
+	}
+
+	protected Encoder<?> getJackson2JsonEncoder() {
+		return (this.jackson2JsonEncoder != null ? this.jackson2JsonEncoder : new Jackson2JsonEncoder());
+	}
+
+}
+
